requirements.txt
django~=3.0.0
requests
beautifulsoup4
python-telegram-bot
django-bootstrap4